액션 메소드 메뉴 권한 검사

+++ src/main/java/fund/controller/CodeController.java
@@ -8,7 +8,9 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import fund.dto.Code;
 import fund.mapper.CodeMapper;
+import fund.service.C;
 import fund.service.LogService;
+import fund.service.UserService;
 
 @Controller
 public class CodeController extends BaseController{
@@ -18,6 +20,7 @@ public class CodeController extends BaseController{
 
     @RequestMapping("/code/list.do")
     public String list(Model model, @RequestParam("gid") int codeGroupId) {
+        if (!UserService.canAccess(C.메뉴_기초정보관리)) return "redirect:/home/logout.do";
         model.addAttribute("list", codeMapper.selectByCodeGroupId(codeGroupId));
         addCodesToModel(model, codeGroupId);
         return "code/list";
@@ -25,6 +28,7 @@ public class CodeController extends BaseController{
 
     @RequestMapping(value="/code/create.do", method=RequestMethod.GET)
     public String create(Model model, @RequestParam("gid") int codeGroupId) {
+        if (!UserService.canAccess(C.메뉴_기초정보관리)) return "redirect:/home/logout.do";
         Code code = new Code();
         code.setCodeGroupId(codeGroupId);
         model.addAttribute("code", code);
@@ -39,6 +43,7 @@ public class CodeController extends BaseController{
     @RequestMapping(value="/code/create.do", method=RequestMethod.POST)
     public String create(Model model, Code code, @RequestParam("gid") int codeGroupId) {
         try {
+            if (!UserService.canAccess(C.메뉴_기초정보관리)) return "redirect:/home/logout.do";
             codeMapper.insert(code);
             return "redirect:list.do?gid=" + codeGroupId;
         } catch (Exception e) {
@@ -49,6 +54,7 @@ public class CodeController extends BaseController{
 
     @RequestMapping(value="/code/edit.do", method=RequestMethod.GET)
     public String edit(Model model, @RequestParam("id") int id, @RequestParam("gid") int codeGroupId) {
+        if (!UserService.canAccess(C.메뉴_기초정보관리)) return "redirect:/home/logout.do";
         model.addAttribute("code", codeMapper.selectById(id));
         addCodesToModel(model, codeGroupId);
         return "code/edit";
@@ -57,6 +63,7 @@ public class CodeController extends BaseController{
     @RequestMapping(value="/code/edit.do", method=RequestMethod.POST, params="cmd=save")
     public String edit(Model model, Code code, @RequestParam("gid") int codeGroupId) {
         try {
+            if (!UserService.canAccess(C.메뉴_기초정보관리)) return "redirect:/home/logout.do";
             codeMapper.update(code);
             return "redirect:list.do?gid=" + codeGroupId;
         } catch (Exception e) {
@@ -68,6 +75,7 @@ public class CodeController extends BaseController{
     @RequestMapping(value="/code/edit.do", method=RequestMethod.POST, params="cmd=delete")
     public String delete(Model model, @RequestParam("id") int id, @RequestParam("gid") int codeGroupId) {
         try {
+            if (!UserService.canAccess(C.메뉴_기초정보관리)) return "redirect:/home/logout.do";
             codeMapper.delete(id);
             return "redirect:list.do?gid=" + codeGroupId;
         } catch (Exception e) {
